Styling for card, put the create-card display into the card itself. Added some nice temp buttons. confirm dialogue for remove.

+++ app/controllers/cardview.js
@@ -2,6 +2,28 @@ import Ember from 'ember';
 
 export default Ember.Controller.extend({
   actions: {
+
+    openCardCreator() {
+      Ember.$('.createStart').hide();
+      Ember.$('.slotDisplay').hide();
+      Ember.$('.createSubmit').css("display","inline-block");
+      Ember.$('.createCancel').css("display","inline-block");
+      Ember.$('.slotCreate').css("display","table-cell");
+    },
+
+    cancelNewCard() {
+      this.set('title', '');
+      this.set('year', '');
+      this.set('abstract', '');
+      this.set('isRead', false);
+
+      Ember.$('.createStart').css("display","inline-block");
+      Ember.$('.slotDisplay').css("display","table-cell");
+      Ember.$('.createSubmit').hide();
+      Ember.$('.slotCreate').hide();
+      Ember.$('.createCancel').hide();
+    },
+
     createNewCard() {
       var card = this.store.createRecord("card", {
         title: this.get('title'),
@@ -15,22 +37,32 @@ export default Ember.Controller.extend({
       this.set('year', '');
       this.set('abstract', '');
       this.set('isRead', false);
+
+      Ember.$('.createStart').css("display","inline-block");
+      Ember.$('.slotDisplay').css("display","table-cell");
+      Ember.$('.createSubmit').hide();
+      Ember.$('.slotCreate').hide();
+      Ember.$('.createCancel').hide();
     },
 
     deleteCard(card) {
-      card.deleteRecord();
-      card.save();
+      if(confirm("Are you sure you'd like to delete this?")) {
+        card.deleteRecord();
+        card.save();
+      } else {
+        return false;
+      }
     },
 
-    toggleCardRead(card) {
-        var isitRead =  card.get('isRead');
-        if (isitRead === false) {
-          card.set('isRead',true);
-        } else {
-          card.set('isRead',false);
-        }
-        card.save();
-    },
+    // toggleCardRead(card) {
+    //     var isitRead =  card.get('isRead');
+    //     if (isitRead === false) {
+    //       card.set('isRead',true);
+    //     } else {
+    //       card.set('isRead',false);
+    //     }
+    //     card.save();
+    // },
 
     viewCard(card) {
       var title = card.get('title');
